[FLINK-25372] Move ThreadDumpInfo out from taskmanager-specific package
